✅ test(paiir): realign paiir coverage around add pipeline
