+++ tests/paiir/pipeline/avgpool/test_compensation.py
