+++ tests/paiir/pipeline/avgpool/test_transient.py
